+++ src/Router.tsx
@@ -1,4 +1,5 @@
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import React from 'react';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Login from './pages/Login/Login';
 import Todo from './pages/Todo/Todo';
 
@@ -14,3 +15,9 @@ const Router = () => {
 };
 
 export default Router;
+
+// ts / tsx
+// 타입을 깔았기 때문에 react 도 명시를 해달라는 것
+// 타입스크립트가 의도치 않은 일을 발생시키지 않게 하기 위함
+// 번들링 과정에서도
+// tsx는 컴포넌트라 따로 번들링 해줘야 함
