--- src/index.js
@@ -1,5 +1,0 @@
-import ReactDOM from 'react-dom/client';
-import Router from './Router';
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(<Router />);
